pyproject.toml
[project]
name = "mlobstaclecourse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "glfw==2.7.0",
    "multipledispatch==1.0.0",
    "numpy==1.26.4",
    "pyglm==2.7.1",
    "pyopengl==3.1.7",
]

[project.optional-dependencies]
cpu = ["torch>=2.7.0", "torchvision>=0.22.0"]
cu128 = ["torch>=2.7.0", "torchvision>=0.22.0"]

[tool.uv]
conflicts = [[{ extra = "cpu" }, { extra = "cu128" }]]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu128", extra = "cu128" },
]
torchvision = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu128", extra = "cu128" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[dependency-groups]
dev = [
    "ruff>=0.12.11",
]
